\name{aaa-class}
\Rdversion{1.1}
\docType{class}
\alias{aaa-class}
\title{Class \code{"aaa"}}
\description{
Class \code{aaa} is for elements of the free antiassociative algebra
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("aaa", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{single_indeterminate_name1}:}{Object of class \code{"character"}}
    \item{\code{single_indeterminate_coeff}:}{Object of class \code{"numeric"}}
    \item{\code{double_indeterminate_name1}:}{Object of class \code{"character"}}
    \item{\code{double_indeterminate_name2}:}{Object of class \code{"character"}}
    \item{\code{double_indeterminate_coeff}:}{Object of class \code{"numeric"}}
    \item{\code{triple_indeterminate_name1}:}{Object of class \code{"character"}}
    \item{\code{triple_indeterminate_name2}:}{Object of class \code{"character"}}
    \item{\code{triple_indeterminate_name3}:}{Object of class \code{"character"}}
    \item{\code{triple_indeterminate_coeff}:}{Object of class \code{"numeric"}}
  }
}
\author{Robin K. S. Hankin}
\examples{
showClass("aaa")
}
\keyword{classes}
