\name{evitaicossa-package}
\alias{evitaicossa-package}
\alias{evitaicossa}
\docType{package}
\title{
\packageTitle{evitaicossa}
}
\description{
\packageDescription{evitaicossa}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{evitaicossa}
\packageIndices{evitaicossa}

Functionality to work with the free antiassociative algebra in R.  The
hex sticker features an image taken from
\insertCite{hoffnung1959;textual}{evitaicossa} in which musical concepts
[pizzicato, crescendo, etc] are given whimsical visual form.  The
character on the hex sticker is captioned \dQuote{A Discord}: Hoffnung's
interpretation of the musical concept of dissonance.  In the book, the
preceding image was a \dQuote{\emph{chord}}, evoking harmony.  The
discord, on the other hand, embodies--for me at
least--antiassociativity: everything is wrong, wrong, wrong.

}
\author{
\packageAuthor{evitaicossa}

Maintainer: \packageMaintainer{evitaicossa}
}
\keyword{ package }

\references{\insertAllCited{}}

\seealso{\code{\link{aaa}}}

\examples{
x <- raaa()
x
y <- raaa()

x+y
x*y

}
