\name{jitter}
\alias{jitter}
\alias{jitter-methods}
\alias{jitter,aaa-method}
\alias{jitter,ANY-method}
\alias{jitter,disord-method}
\alias{jitter,aaa-method}
\title{Jittering aaa objects}
\description{Jitters the components of an aaa object.}
\value{Returns an object of class \code{aaa}}
\author{Robin K. S. Hankin}
\examples{
jitter(raaa())
jitter(raaa(), amount=1e-6)
}
