\name{raaa}
\alias{raaa}
\alias{raaaa}
\title{Random elements of the free antiassociative algebra}
\description{
Random elements of the free antiassociative algebra, intended as quick
\dQuote{get you going} examples of \code{aaa} objects
}
\usage{
raaa(n = 4, s = 3)
raaaa(n = 10, s = 30)
}

\arguments{
  \item{n}{Number of terms to generate}
  \item{s}{Number of symbols to use in the alphabet}
}
\details{

Function \code{raaa()} returns a random  \code{aaa} object.
Function \code{raaaa()} returns, by default,
a more complicated \code{aaa} object.
}
\value{Returns an object of class \code{aaa}}
\author{Robin K. S. Hankin}
\examples{
raaa()
raaaa()
}
