\name{show}
\alias{show}
\alias{print}
\alias{print.aaa}
\alias{aaa_show}
\title{Print method for antiassociative algebra objects}
\description{Show methods for \code{aaa} objects}
\usage{
\S4method{show}{aaa}(object)
aaa_show(a)
}
\arguments{
  \item{a,object}{Object of class \code{aaa}}
}
\details{

A bunch of functionality to print \code{aaa} objects
  
}
  
\author{Robin K. S. Hankin}
\examples{
1+1
}

